refactor: extract config types to src/config/types.ts (Phase 2.3)
- Create src/config/types.ts with all config interfaces (333 lines)
- Create src/config/index.ts for module re-exports (41 lines)
- Update src/config.ts to import types from new location
- Remove ~300 lines of type definitions from config.ts (855→555 lines)
- Update REFACTORING-PLAN.md with Phase 2.3 progress

Extracted types:
- WorkspaceConfig - Workspace configuration interface
- ResolvedConfig - Resolved configuration with all values set
- ApprovedPatternConfig - Approved bash command patterns
- ApprovedPathPatternConfig - Approved file path patterns
- ToolsConfig - Per-tool configuration
